+++ views/templates/content/pages/editor.php
@@ -8,7 +8,7 @@ use Flextype\Component\{I18n\I18n, Registry\Registry, Html\Html, Form\Form, Http
     Themes::view('admin/views/partials/navbar')
         ->assign('links',   ['pages' => ['url' => Http::getBaseUrl() . '/admin/pages/edit?page=' . $page_name, 'title' => I18n::find('admin_pages_edit_page', Registry::get('system.locale'))]])
         ->assign('buttons', ['pages' =>
-                                        ['url' => Http::getBaseUrl() . '/admin/pages/edit?page=' . $page_name . '&expert=true', 'title' => 'Switch to expert mode']])
+                                        ['url' => Http::getBaseUrl() . '/admin/pages/edit?page=' . $page_name . '&expert=true', 'title' => I18n::find('admin_pages_switch_to_expert_mode', Registry::get('system.locale'))]])
         ->display();
     Themes::view('admin/views/partials/content-start')->display();
 ?>
@@ -52,7 +52,7 @@ use Flextype\Component\{I18n\I18n, Registry\Registry, Html\Html, Form\Form, Http
                 <?php
                     echo (
                         Form::label('page_template', I18n::find('admin_pages_template', Registry::get('system.locale')),  ['for' => 'pageTemplate']).
-                        Form::select('page_template', ['default' => 'default'], 'default', ['class' => 'form-control', 'id' => 'pageTemplate'])
+                        Form::select('page_template', $templates, $page_template, ['class' => 'form-control', 'id' => 'pageTemplate'])
                     );
                 ?>
             </div>
@@ -94,7 +94,7 @@ use Flextype\Component\{I18n\I18n, Registry\Registry, Html\Html, Form\Form, Http
                         <input type="file" name="file">
                         <?php
                             echo (
-                                Form::submit('upload_file', 'Upload', array('class' => '')).
+                                Form::submit('upload_file', I18n::find('admin_pages_files_upload', Registry::get('system.locale')), array('class' => '')).
                                 Form::close()
                             )
                         ?>