+++ CHANGELOG.md
@@ -1,2 +1,15 @@
+# v0.1.1, 2018-11-17
+* Page editor: fix issue with getting theme templates
+* System Settings: added ability to select locales from list
+* Forum link removed
+* English translates updates
+* Page editor: fix issue with locales
+* Russian translates added
+* German translates added - Thanks to Andreas Lenhardt
+* Ukraine translates added - Thanks to Shakhrom Mukumov
+* Clear Cache functionality added
+* Clear Cache on plugin activation/deactivation
+* Plugins Manager: don't allow to deactivate admin plugin
+
 # v0.1.0, 2018-11-16
 * Initial release